take function out of MainActivity
